Emilia-main/requirements.txt
Pyrogram       ==2.0.106
pyromod        ==3.1.6
shazamio       ==0.6.0
captcha        ==0.1.1
python-decouple==3.8
py-tgcalls     ==2.0.2
python-dotenv  ==1.0.1
requests       ==2.31.0
pytube         ==15.0.0
gTTS           ==2.5.1
pymongo        ==4.7.2